Align resource loading logic in `AntClassLoader` with upstream
