+++ core/src/main/java/jenkins/ClassLoaderReflectionToolkit.java
@@ -9,7 +9,6 @@ import java.util.Enumeration;
 import edu.umd.cs.findbugs.annotations.CheckForNull;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import jenkins.util.AntClassLoader;
-import jenkins.util.AntWithFindResourceClassLoader;
 
 /**
  * Reflective access to various {@link ClassLoader} methods which are otherwise {@code protected}.
@@ -47,8 +46,8 @@ public class ClassLoaderReflectionToolkit {
     }
 
     private static Object getClassLoadingLock(ClassLoader cl, String name) {
-        if (cl instanceof AntWithFindResourceClassLoader) {
-            return ((AntWithFindResourceClassLoader) cl).getClassLoadingLock(name);
+        if (cl instanceof AntClassLoader) {
+            return ((AntClassLoader) cl).getClassLoadingLock(name);
         }
         return invoke(GetClassLoadingLock.GET_CLASS_LOADING_LOCK, RuntimeException.class, cl, name);
     }
@@ -75,8 +74,8 @@ public class ClassLoaderReflectionToolkit {
     public static @CheckForNull Class<?> _findLoadedClass(ClassLoader cl, String name) {
         synchronized (getClassLoadingLock(cl, name)) {
             Class<?> c;
-            if (cl instanceof AntWithFindResourceClassLoader) {
-                c = ((AntWithFindResourceClassLoader) cl).findLoadedClass2(name);
+            if (cl instanceof AntClassLoader) {
+                c = ((AntClassLoader) cl).findLoadedClass2(name);
             } else {
                 c = (Class) invoke(FindLoadedClass.FIND_LOADED_CLASS, RuntimeException.class, cl, name);
             }
@@ -132,8 +131,8 @@ public class ClassLoaderReflectionToolkit {
      */
     public static @CheckForNull URL _findResource(ClassLoader cl, String name) {
         URL url;
-        if (cl instanceof AntWithFindResourceClassLoader) {
-            url = ((AntWithFindResourceClassLoader) cl).findResource(name);
+        if (cl instanceof AntClassLoader) {
+            url = ((AntClassLoader) cl).findResource(name);
         } else if (cl instanceof URLClassLoader) {
             url = ((URLClassLoader) cl).findResource(name);
         } else {
@@ -162,8 +161,8 @@ public class ClassLoaderReflectionToolkit {
      */
     public static @NonNull Enumeration<URL> _findResources(ClassLoader cl, String name) throws IOException {
         Enumeration<URL> urls;
-        if (cl instanceof AntWithFindResourceClassLoader) {
-            urls = ((AntWithFindResourceClassLoader) cl).findResources(name);
+        if (cl instanceof AntClassLoader) {
+            urls = ((AntClassLoader) cl).findResources(name);
         } else {
             urls = (Enumeration<URL>) invoke(FindResources.FIND_RESOURCES, IOException.class, cl, name);
         }
