+++ core/src/main/java/jenkins/util/AntClassLoader.java
@@ -35,7 +35,6 @@ import org.apache.tools.zip.ZipLong;
 import org.kohsuke.accmod.Restricted;
 import org.kohsuke.accmod.restrictions.NoExternalUse;
 
-import edu.umd.cs.findbugs.annotations.CheckForNull;
 import java.io.ByteArrayOutputStream;
 import java.io.Closeable;
 import java.io.File;
@@ -48,6 +47,7 @@ import java.nio.file.Files;
 import java.security.CodeSource;
 import java.security.ProtectionDomain;
 import java.security.cert.Certificate;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.Enumeration;
 import java.util.HashMap;
@@ -284,7 +284,6 @@ public class AntClassLoader extends ClassLoader implements SubBuildListener, Clo
      * @param classpath The classpath to use to load classes.
      */
     public AntClassLoader(final ClassLoader parent, final Project project, final Path classpath) {
-        super(parent);  // KK patch for JENKINS-21579
         setParent(parent);
         setClassPath(classpath);
         setProject(project);
@@ -371,7 +370,6 @@ public class AntClassLoader extends ClassLoader implements SubBuildListener, Clo
      *                    load the a class through this loader.
      */
     public AntClassLoader(final ClassLoader parent, final boolean parentFirst) {
-        super(parent);  // KK patch for JENKINS-21579
         setParent(parent);
         project = null;
         this.parentFirst = parentFirst;
@@ -512,6 +510,12 @@ public class AntClassLoader extends ClassLoader implements SubBuildListener, Clo
             return;
         }
         pathComponents.addElement(file);
+    }
+
+    public void addPathFiles(Collection<File> paths) throws IOException {
+        for (File f : paths) {
+            addPathFile(f);
+        }
     }
 
     /**
@@ -902,9 +906,7 @@ public class AntClassLoader extends ClassLoader implements SubBuildListener, Clo
         if (url != null) {
             log("Resource " + name + " loaded from parent loader", Project.MSG_DEBUG);
         } else {
-            // try and load from this loader if the parent either didn't find
-            // it or wasn't consulted.
-            url = getUrl(pathComponents, name);
+            url = getUrl(name);
         }
         if (url == null && !isParentFirst(name)) {
             // this loader was first but it didn't find it - try the parent
@@ -924,15 +926,17 @@ public class AntClassLoader extends ClassLoader implements SubBuildListener, Clo
     }
 
     /**
-     * Finds a matching file by iterating pathComponents.
+     * Finds a matching file by iterating through path components.
      *
-     * @param pathComponents Path to a folder, split into the individual folder names
      * @param name File to find
-     * @return Url to found object
+     * @return A {@code URL} object for reading the resource, or {@code null} if the
+     *     resource could not be found
      */
-    @CheckForNull
-    protected URL getUrl(Iterable<File> pathComponents, String name) {
+    private URL getUrl(String name) {
         URL url = null;
+
+        // try and load from this loader if the parent either didn't find
+        // it or wasn't consulted.
         for (final File pathComponent : pathComponents) {
             url = getResourceURL(pathComponent, name);
             if (url != null) {
@@ -940,6 +944,7 @@ public class AntClassLoader extends ClassLoader implements SubBuildListener, Clo
                 break;
             }
         }
+
         return url;
     }
 
@@ -947,9 +952,6 @@ public class AntClassLoader extends ClassLoader implements SubBuildListener, Clo
      * Finds all the resources with the given name. A resource is some
      * data (images, audio, text, etc) that can be accessed by class
      * code in a way that is independent of the location of the code.
-     *
-     * <p>Would override getResources if that wasn't final in Java
-     * 1.4.</p>
      *
      * @param name name of the resource
      * @return possible URLs as enumeration
@@ -960,6 +962,18 @@ public class AntClassLoader extends ClassLoader implements SubBuildListener, Clo
     public Enumeration<URL> getNamedResources(final String name)
         throws IOException {
         return findResources(name, false);
+    }
+
+    /**
+     * Finds the resource with the given name.
+     *
+     * @param name The resource name
+     * @return A {@code URL} object for reading the resource, or {@code null} if the
+     *     resource could not be found
+     */
+    @Override
+    public URL findResource(final String name) {
+        return getUrl(name);
     }
 
     /**
@@ -982,26 +996,21 @@ public class AntClassLoader extends ClassLoader implements SubBuildListener, Clo
      *
      * @param name The resource name to search for.
      *             Must not be {@code null}.
-     * @param parentHasBeenSearched whether ClassLoader.this.parent
-     * has been searched - will be true if the method is (indirectly)
-     * called from ClassLoader.getResources
+     * @param skipParent whether to skip searching the parent first - will be false if the method is
+     *     invoked from {@link #getResources(String)} or {@link #getNamedResources(String)} and true
+     *     if the method is invoked from {@link #findResources(String)}.
      * @return an enumeration of URLs for the resources
      * @exception IOException if I/O errors occurs (can't happen)
      */
     protected Enumeration<URL> findResources(final String name,
-                                             final boolean parentHasBeenSearched)
+                                             final boolean skipParent)
         throws IOException {
         final Enumeration<URL> mine = new ResourceEnumeration(name);
         Enumeration<URL> base;
-        if (parent != null && (!parentHasBeenSearched || parent != getParent())) {
+        if (parent != null && !skipParent) {
             // Delegate to the parent:
             base = parent.getResources(name);
-            // Note: could cause overlaps in case
-            // ClassLoader.this.parent has matches and
-            // parentHasBeenSearched is true
         } else {
-            // ClassLoader.this.parent is already delegated to for example from
-            // ClassLoader.getResources, no need:
             base = Collections.emptyEnumeration();
         }
         if (isParentFirst(name)) {
@@ -1578,6 +1587,27 @@ public class AntClassLoader extends ClassLoader implements SubBuildListener, Clo
     @Override
     public String toString() {
         return "AntClassLoader[" + getClasspath() + "]";
+    }
+
+    /**
+     * Public version of {@link ClassLoader#findLoadedClass(String)}
+     */
+    public Class<?> findLoadedClass2(String name) {
+        return findLoadedClass(name);
+    }
+
+    /**
+     * Public version of {@link ClassLoader#getClassLoadingLock(String)}
+     */
+    @Override
+    public Object getClassLoadingLock(String className) {
+        return super.getClassLoadingLock(className);
+    }
+
+    /** {@inheritDoc} */
+    @Override
+    public Enumeration<URL> getResources(String name) throws IOException {
+        return getNamedResources(name);
     }
 
     /** {@inheritDoc} */
